backend/pyproject.toml
[project]
name = "gemini-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a>=0.44",
    "a2a-sdk>=0.2.6",
    "fastapi>=0.115.12",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-aiplatform>=1.96.0",
    "google-genai>=1.19.0",
    "ipykernel>=6.29.5",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "loguru>=0.7.3",
    "pillow>=11.2.1",
    "python-a2a>=0.5.9",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
